requirements.txt
langchain
langchain_community
langchain_groq
python-dotenv
langchain_core
crewai
streamlit
crewai-tools